refactor: processDustSetData 함수 분리 및 type 분리
