+++ app/api/walkingIndex.ts
